Ajout des enums Speed, et modif des types d'arguments les appelant (float -> double )

+++ software/pc/src/robot/Locomotion.java
@@ -1015,7 +1015,7 @@ public class Locomotion implements Service
         }
     }
 
-    public void setRotationnalSpeed(float rotationSpeed) throws SerialConnexionException
+    public void setRotationnalSpeed(double rotationSpeed) throws SerialConnexionException
     {
         deplacements.setRotationnalSpeed(rotationSpeed);
     }
